requirements.txt
uvicorn==0.32.0
fastapi==0.115.4
PyJWT==2.9.0